requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
chardet==3.0.4
discord.py==1.3.3
google==2.0.3
idna==2.9
idna-ssl==1.1.0
multidict==4.7.5
postgis==1.0.4
psycopg2-binary==2.8.6
soupsieve==2.0
typing-extensions==3.7.4.3
websockets==8.1
yarl==1.4.2
